pyproject.toml
[project]
name = "project-banking-widget"
version = "0.1.0"
description = ""
authors = [
    {name = "Artem Kabritskii",email = "artemkabr7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
pre-commit = "^4.2.0"
mypy = "^1.15.0"
isort = "^6.0.1"
pytest-cov = "^6.1.1"

[tool.black]
line-length = 119
target-version = ["py39"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["my_project"]  # замени на имя твоего пакета
src_paths = ["src"]

[tool.mypy]
python_version = 3.9
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
show_error_codes = true
pretty = true
exclude = 'venv|tests'





